Reference

Contents

Reference

Contents

Index

Index

ForestMensuration.β0

ForestMensuration.CubingMethod

ForestMensuration.FittedLinearModel

ForestMensuration.FittedLinearModel

ForestMensuration.Huber

ForestMensuration.MixTerm

ForestMensuration.Newton

ForestMensuration.Smalian

ForestMensuration.StratifiedReport

Base.show

ForestMensuration._amplitude

ForestMensuration._class_breadth

ForestMensuration._class_center

ForestMensuration._create_matrix_formulas!

ForestMensuration._criteria_parameters

ForestMensuration._dependent_variable

ForestMensuration._finite_population

ForestMensuration._finite_sample

ForestMensuration._fit_regression!

ForestMensuration._indepedent_variable

ForestMensuration._infinite_population

ForestMensuration._infinite_sample

ForestMensuration._inventory_report

ForestMensuration._predict

ForestMensuration._simple_frequency

ForestMensuration._sturges

ForestMensuration.adjr2

ForestMensuration.artificial_form_factor

ForestMensuration.bark_factor

ForestMensuration.bole_volume

ForestMensuration.coef_table

ForestMensuration.coefnames

ForestMensuration.cone_volume

ForestMensuration.confidence_interval

ForestMensuration.criteria_table

ForestMensuration.criteria_table

ForestMensuration.cubage

ForestMensuration.cubage

ForestMensuration.cubage

ForestMensuration.cubage

ForestMensuration.cylinder_volume

ForestMensuration.diametric_table

ForestMensuration.diametric_table

ForestMensuration.diametric_table

ForestMensuration.diametric_table

ForestMensuration.dispersion

ForestMensuration.frequency_table

ForestMensuration.frequency_table

ForestMensuration.frequency_table

ForestMensuration.frequency_table

ForestMensuration.homoscedasticity

ForestMensuration.loglikelihood

ForestMensuration.n_coef

ForestMensuration.natural_form_factor

ForestMensuration.normality

ForestMensuration.p_result

ForestMensuration.predict

ForestMensuration.predict

ForestMensuration.predict

ForestMensuration.quotient_form

ForestMensuration.r2

ForestMensuration.regression

ForestMensuration.residuals

ForestMensuration.simple_casual_sampling

ForestMensuration.simple_casual_sampling

ForestMensuration.stderror

ForestMensuration.syx

ForestMensuration.syx_in_percentage

StatsAPI.aic

StatsAPI.aicc

StatsAPI.bic

StatsAPI.coef

StatsAPI.deviance

StatsAPI.dof

StatsAPI.dof_residual

StatsAPI.modelmatrix

StatsAPI.nobs

StatsAPI.nulldeviance

StatsAPI.nulldeviance

ForestMensuration.MixTerm — Type

Union type representing a mixed term, which can be a single AbstractTerm or a tuple of AbstractTerms.

ForestMensuration.β0 — Constant

Represents an intercept term for linear models.

ForestMensuration.CubingMethod — Type

Abstract type representing a method for cubing (calculating volume).

ForestMensuration.FittedLinearModel — Type

Represents a fitted linear model.

Fields

formula::F: The formula used for the model.

data::D: The data frame containing the data.

β::Array{T, 1}: The regression coefficients.

σ²::T: The variance of residuals.

RMSE::T: The root mean squared error.

chol::C: The Cholesky decomposition of X'X.

ForestMensuration.FittedLinearModel — Method

Creates a new FittedLinearModel instance.

Arguments

formula::F: The formula used for the model.

data::D: The data frame containing the data.

β::Array{T, 1}: The regression coefficients.

σ²::T: The variance of residuals.

RMSE::T: The root mean squared error.

chol::C: The Cholesky decomposition of X'X.

Returns

FittedLinearModel{F, D, T, C}: A new fitted linear model.

ForestMensuration.Huber — Type

Huber Method: The Huber method measures the diameter or circumference at the midpoint of the section, and the volume is determined by:

v = v0 + Σi=1:n(vi) + vt

vi = gi * li

Where:

v0 = volume of the stump;

vi = volume of intermediate sections;

vt = volume of the cone;

g = basal area;

l = length.

ForestMensuration.Newton — Type

Newton Method: The Newton method involves measuring at 3 positions along each section (at the ends and in the middle of the logs). Therefore, it is a more laborious method than the others, but the estimated volume will be more accurate.

v = v0 + Σi=1:n(vi) + vt

vi = (gi + gm + gi+1)/2 * li

Where:

v0 = volume of the stump;

vi = volume of intermediate sections;

vt = volume of the cone;

g = basal area;

gm = basal area at the midpoint of the section;

l = length.

ForestMensuration.Smalian — Type

Smalian Method: The Smalian method measures diameters or circumferences at the ends of each section and calculates the total volume by:

Vt = v0 + Σi=1:n(vi) + vt

v0 = g0 * l0

vi = (gi + gi+1)/2 * li

vt = (1/3) * gn * ln

Where:

v0 = volume of the stump;

vi = volume of intermediate sections;

vt = volume of the cone;

g = basal area;

l = length.

ForestMensuration.StratifiedReport — Type

Represents a report for stratified sampling in forest inventory.

Fields

anova::AbstractDataFrame: ANOVA table.

auxiliary_table::AbstractDataFrame: Auxiliary table with additional statistics.

result_table::AbstractDataFrame: Result table with final estimates.

Base.show — Method

Display the equation of the fitted linear model.

Arguments

io::IO: The IO stream to which the output is written.

fitted_model::FittedLinearModel: The fitted linear model to be displayed.

ForestMensuration._amplitude — Method

Calculates the amplitude (range) of a vector of values.

Arguments

x::Vector: The vector of values.

Returns

Real: The amplitude (range).

ForestMensuration._class_breadth — Method

Calculates the class breadth (width) for a given amplitude and number of classes.

Arguments

h::Real: The amplitude.

k::Int: The number of classes.

Returns

Real: The class breadth (width).

ForestMensuration._class_center — Method

Calculates the class center for a given value and class width.

Arguments

x::Real: The value.

hi::Real: The class width.

Returns

Real: The class center.

ForestMensuration._create_matrix_formulas! — Method

Create matrix formulas for linear regression models.

Arguments

matrix_formulas::Vector{MatrixTerm}: Vector to store the resulting matrix formulas.

x_term_list::Vector{MixTerm}: List of independent variable terms.

q_term::AbstractTerm...: Additional terms for the model.

effect::Symbol: Type of effect to consider (:additive or :interactive).

Returns

Void: Modifies matrix_formulas in place.

ForestMensuration._criteria_parameters — Method

Calculate various criteria parameters for a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Matrix{Float64}: The matrix of criteria parameters.

ForestMensuration._dependent_variable — Method

Generates a list of transformed dependent variable terms.

Arguments

y_term::AbstractTerm: The dependent variable term.

x_term::AbstractTerm: The independent variable term.

Returns

Vector{AbstractTerm}: A list of transformed dependent variable terms.

ForestMensuration._finite_population — Method

Calculates the required sample size for a finite population using an iterative approach with a maximum iteration limit to prevent infinite loops.

Arguments

cv::Real: Coefficient of variation.

ni::Real: Initial sample size.

N::Int: Population size.

e::Real: Desired error margin.

α::Real: Confidence level.

max_iterations::Int=1000: Maximum number of iterations (default is 1000).

Returns

Int64: The required sample size for a finite population.

ForestMensuration._finite_sample — Method

Calculates the sample size for a finite population.

Arguments

t::Real: T-value for the given confidence level.

cv::Real: Coefficient of variation.

e::Real: Desired error margin.

N::Int64: Population size.

Returns

Float64: The sample size for a finite population.

ForestMensuration._fit_regression! — Method

Fit linear regression models to the data.

Arguments

fitted_models::Vector{FittedLinearModel}: Vector to store the fitted models.

y_term_list::Vector{AbstractTerm}: List of dependent variable terms.

matrix_formulas::Vector{MatrixTerm}: List of matrix formulas for the models.

data::AbstractDataFrame: The data frame containing the data.

y_observed::Vector{<:Real}: Observed values of the dependent variable.

x_observed::Vector{<:Real}: Observed values of the independent variable.

Returns

Void: Modifies fitted_models in place.

ForestMensuration._indepedent_variable — Method

Generates a list of transformed independent variable terms and their interactions.

Arguments

x_term::AbstractTerm: The independent variable term.

Returns

Vector{MixTerm}: A list of transformed independent variable terms and their interactions.

ForestMensuration._infinite_population — Method

Calculates the required sample size for an infinite population using an iterative approach with a maximum iteration limit to prevent infinite loops.

Arguments

cv::Real: Coefficient of variation.

ni::Real: Initial sample size.

e::Real: Desired error margin.

α::Real: Confidence level.

max_iterations::Int=1000: Maximum number of iterations (default is 1000).

Returns

Int64: The required sample size for an infinite population.

ForestMensuration._infinite_sample — Method

Calculates the sample size for an infinite population.

Arguments

t::Real: T-value for the given confidence level.

cv::Real: Coefficient of variation.

e::Real: Desired error margin.

Returns

Float64: The sample size for an infinite population.

ForestMensuration._inventory_report — Method

Generate an inventory report.

Arguments

x̅::F: Mean volume per plot.

cv::F: Coefficient of variation.

Ttab::F: T-value for the confidence level.

s²x̅::F: Mean variance.

sx̅::F: Standard error.

plot_area::Union{I, F, S}: Area of the plot.

f::F: Sampling fraction.

population::String: Population type ("infinite" or "finite").

n::Union{I, T}: Number of measured plots.

required_plots::Union{I, T}: Number of required plots.

missing_plots::Union{I, T}: Number of missing plots.

N::I: Total number of plots.

lg::Symbol: Language for the report (":pt" for Portuguese, ":en" for English).

Returns

DataFrame: The inventory report.

ForestMensuration._predict — Method

Predict values based on different transformation functions.

Arguments

ŷ::Vector{<:Real}: Predicted values.

x::Vector{<:Real}: Independent variable values.

σ²::Real: Variance of residuals.

function_name::Symbol: Name of the transformation function.

Returns

Vector{<:Real}: Transformed predicted values.

ForestMensuration._simple_frequency — Method

Calculates the simple frequency of unique values in a vector.

Arguments

x::Vector: The vector of values.

Returns

Vector: The frequency of each unique value.

ForestMensuration._sturges — Method

Calculates the number of classes using Sturges' formula.

Arguments

n::Int: The number of observations.

Returns

Int: The number of classes.

ForestMensuration.adjr2 — Method

Calculate the adjusted R-squared value for a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Float64: The adjusted R-squared value.

ForestMensuration.artificial_form_factor — Method

Artificial Form Factor (aff): For the calculation of the artificial form factor, the volume of the reference cylinder will have a diameter equal to the tree's DBH.

aff = Rigorous Vol / Cylinder Vol 1.3

Where:

Rigorous Vol = total volume determined by one of the methods: Smalian, Huber, or Newton;

Cylinder Vol 1.3 = volume of a cylinder with height and diameter equal to the total height and DBH of the tree.

Arguments

vt::Real: The total volume of the tree.

ht::Real: The total height of the tree.

dbh::Real: The diameter at breast height of the tree.

Returns

Float64: The artificial form factor.

ForestMensuration.bark_factor — Method

Calculates the bark factor, used to estimate the volume without bark.

The bark factor is used to estimate the volume without bark by considering the ratio of bark thickness to total diameter.

Arguments

d::Vector{<:Real}: Vector of diameters.

e::Vector{<:Real}: Vector of bark thicknesses in millimeters.

Returns

Float64: The bark factor.

Example

Calculate bark factor

k = bark_factor(d, e)

Calculate total volume without bark

totalvolumewithoutbark = totalvolume * k^2

ForestMensuration.bole_volume — Method

Calculate tree bole volume cubic meters using Smalian, Newton, or Huber methods. The methods involve dividing the tree trunk into n sections (logs). In each section, diameters and lengths are measured at positions that vary according to the technique employed.

Arguments

method::Type{<:CubingMethod}: The method used for cubing (Smalian, Huber, or Newton).

h::Vector{<:Real}: Vector of heights.

d::Vector{<:Real}: Vector of diameters.

Returns

Float64: The volume of the bole in cubic meters.

ForestMensuration.coef_table — Method

Create a table of coefficients for a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

DataFrame: The coefficient table.

ForestMensuration.coefnames — Method

Get names of the coefficients of a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Vector{String}: The names of the coefficients.

ForestMensuration.cone_volume — Method

Calculates the volume of a cone, used to estimate the final portion (vn) of the tree, typically considered to have a conical shape.

Arguments

h::Real: The height of the cone in meters.

d::Real: The diameter of the cone in centimeters.

Returns

Float64: The volume of the cone in cubic meters.

ForestMensuration.confidence_interval — Method

Calculate the confidence intervals for a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Vector{Float64}: The confidence intervals.

ForestMensuration.criteria_table — Method

Create a criteria table for a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

DataFrame: The criteria table.

ForestMensuration.criteria_table — Method

Create a criteria table for multiple fitted linear models.

Arguments

fitted_model::Vector{<:FittedLinearModel}: The vector of fitted linear models.

Returns

DataFrame: The criteria table.

ForestMensuration.cubage — Method

Calculate tree cubage using grouped data from a DataFrame. The methods involve dividing the tree trunk into n sections (logs). In each section, diameters and lengths are measured at positions that vary according to the technique employed.

Arguments

method::Type{<:CubingMethod}: The method used for cubing (Smalian, Huber, or Newton).

tree::Symbol: The symbol representing the tree identifier.

h::Symbol: The symbol representing the heights in the DataFrame.

d::Symbol: The symbol representing the diameters in the DataFrame.

data::AbstractDataFrame: The DataFrame containing the tree data.

dbh::Float64: Diameter at breast height (default is 1.3 meters).

Returns

DataFrame: A DataFrame with the calculated volumes and form factors for each tree.

ForestMensuration.cubage — Method

Calculate tree cubage including bark factor using grouped data from a DataFrame. The methods involve dividing the tree trunk into n sections (logs). In each section, diameters and lengths are measured at positions that vary according to the technique employed.

Arguments

method::Type{<:CubingMethod}: The method used for cubing (Smalian, Huber, or Newton).

tree::Symbol: The symbol representing the tree identifier.

h::Symbol: The symbol representing the heights in the DataFrame.

d::Symbol: The symbol representing the diameters in the DataFrame.

e::Symbol: The symbol representing the bark thicknesses in the DataFrame.

data::AbstractDataFrame: The DataFrame containing the tree data.

dbh::Float64: Diameter at breast height (default is 1.3 meters).

Returns

DataFrame: A DataFrame with the calculated volumes, form factors, and bark-adjusted volumes for each tree.

ForestMensuration.cubage — Method

Calculate tree cubage including bark factor. The methods involve dividing the tree trunk into n sections (logs). In each section, diameters and lengths are measured at positions that vary according to the technique employed.

Arguments

method::Type{<:CubingMethod}: The method used for cubing (Smalian, Huber, or Newton).

h::Vector{<:Real}: Vector of heights.

d::Vector{<:Real}: Vector of diameters.

e::Vector{<:Real}: Vector of bark thicknesses.

dbh::Float64: Diameter at breast height (default is 1.3 meters).

Returns

DataFrame: A DataFrame with the calculated volumes, form factors, and bark-adjusted volumes.

ForestMensuration.cubage — Method

Calculate tree cubage using Smalian, Newton, or Huber methods. The methods involve dividing the tree trunk into n sections (logs). In each section, diameters and lengths are measured at positions that vary according to the technique employed. Thus, the volume of the sections and the total volume are determined by summing the volume of the sections. Determination can be carried out on felled trees or standing trees using equipment such as the Bitterlich relascope.

Arguments

method::Type{<:CubingMethod}: The method used for cubing (Smalian, Huber, or Newton).

h::Vector{<:Real}: Vector of heights.

d::Vector{<:Real}: Vector of diameters.

dbh::Float64: Diameter at breast height (default is 1.3 meters).

Returns

DataFrame: A DataFrame with the calculated volumes and form factors.

ForestMensuration.cylinder_volume — Method

Calculates the volume of a cylinder, used to estimate the volume (v0) of the tree stump remaining after clear-cutting.

Arguments

h::Real: The height of the cylinder in meters.

d::Real: The diameter of the cylinder in centimeters.

Returns

Float64: The volume of the cylinder in cubic meters.

ForestMensuration.diametric_table — Method

Creates a diametric table for a vector of values given a class width and plot area.

Arguments

x::Vector{<:Real}: The vector of values.

hi::Real: The class width.

plot_area::Real=1.0: The plot area in hectares (default is 1.0).

Returns

DataFrame: A DataFrame containing the diametric table.

ForestMensuration.diametric_table — Method

Creates a diametric table for a vector of values.

Arguments

x::Vector{<:Real}: The vector of values.

plot_area::Real=1.0: The plot area in hectares (default is 1.0).

Returns

DataFrame: A DataFrame containing the diametric table.

ForestMensuration.diametric_table — Method

Creates a diametric table for grouped data in a DataFrame.

Arguments

g::S: The symbol representing the grouping variable.

x::S: The symbol representing the values variable.

data::AbstractDataFrame: The DataFrame containing the data.

plot_area::Real=1.0: The plot area in hectares (default is 1.0).

Returns

DataFrame: A DataFrame containing the diametric table for each group.

ForestMensuration.diametric_table — Method

Creates a diametric table for grouped data in a DataFrame with a specified class width.

Arguments

g::S: The symbol representing the grouping variable.

x::S: The symbol representing the values variable.

hi::Real: The class width.

data::AbstractDataFrame: The DataFrame containing the data.

plot_area::Real=1.0: The plot area in hectares (default is 1.0).

Returns

DataFrame: A DataFrame containing the diametric table for each group.

ForestMensuration.dispersion — Method

Calculate the dispersion of a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Matrix{Float64}: The dispersion matrix.

ForestMensuration.frequency_table — Method

Creates a frequency table for a vector of values given a class width.

Arguments

x::Vector{<:Real}: The vector of values.

hi::Real: The class width.

Returns

DataFrame: A DataFrame containing the frequency table.

ForestMensuration.frequency_table — Method

Creates a frequency table for a vector of values.

Arguments

x::Vector{<:Real}: The vector of values.

Returns

DataFrame: A DataFrame containing the frequency table.

ForestMensuration.frequency_table — Method

Creates a frequency table for grouped data in a DataFrame.

Arguments

g::S: The symbol representing the grouping variable.

x::S: The symbol representing the values variable.

data::AbstractDataFrame: The DataFrame containing the data.

Returns

DataFrame: A DataFrame containing the frequency table for each group.

ForestMensuration.frequency_table — Method

Creates a frequency table for grouped data in a DataFrame with a specified class width.

Arguments

g::S: The symbol representing the grouping variable.

x::S: The symbol representing the values variable.

hi::Real: The class width.

data::AbstractDataFrame: The DataFrame containing the data.

Returns

DataFrame: A DataFrame containing the frequency table for each group.

ForestMensuration.homoscedasticity — Method

Test for homoscedasticity of residuals in a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

WhiteTest: The result of the White test.

ForestMensuration.loglikelihood — Method

Calculate the log-likelihood for a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Float64: The log-likelihood.

ForestMensuration.n_coef — Method

Get the number of coefficients of a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Int: The number of coefficients.

ForestMensuration.natural_form_factor — Method

Natural Form Factor (nff): For the calculation of the natural form factor, the volume of the reference cylinder will have a diameter equal to the diameter taken at 1/10 of the total height.

f0.1h = Rigorous Vol / Cylinder Vol 0.1

Where:

Rigorous Vol = total volume determined by one of the methods: Smalian, Huber, or Newton;

Cylinder Vol 0.1 = volume of a cylinder with height equal to the total height of the tree and diameter taken at 1/10 of the total height.

Interpolate diameter at a given height using linear interpolation.

Arguments

vt::Real: The total volume of the tree.

ht::Real: The total height of the tree.

h::Vector{<:Real}: Vector of heights.

d::Vector{<:Real}: Vector of diameters.

Returns

Float64: The natural form factor.

ForestMensuration.normality — Method

Test for normality of residuals in a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

ExactOneSampleKSTest: The result of the Kolmogorov-Smirnov test.

ForestMensuration.p_result — Method

Check the result of a hypothesis test.

Arguments

test::HypothesisTests.HypothesisTest: The hypothesis test result.

Returns

Bool: True if the p-value is greater than 0.05, otherwise false.

ForestMensuration.predict — Method

Predict values using a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

data::AbstractDataFrame: The data frame containing the new data for prediction.

Returns

Vector{Float64}: The predicted values.

ForestMensuration.predict — Method

Predict a single value using a fitted linear model for a single row of data.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

data::DataFrameRow: A single row of data for prediction.

Returns

Float64: The predicted value.

ForestMensuration.predict — Method

Predict values using the data from the fitted model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Vector{Float64}: The predicted values.

ForestMensuration.quotient_form — Method

Form Quotient (qf): The natural decrease in diameter along the trunk defines the so-called form quotient, which is a ratio between diameters. An example of a form quotient is the Schiffel form quotient, given by:

Q = D(1/2H) / DBH

Where:

Q < 1

D(1/2H) = diameter measured at half the total height of the tree.

Similar to the form factor, the volume of a tree, with or without bark, can be obtained by multiplying the volume of a cylinder by the average form quotient, suitable for the species and the desired volume to be estimated.

Arguments

ht::Real: The total height of the tree.

dbh::Real: The diameter at breast height of the tree.

h::Vector{<:Real}: Vector of heights.

d::Vector{<:Real}: Vector of diameters.

Returns

Float64: The form quotient.

ForestMensuration.r2 — Method

Calculate the R-squared value for a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Float64: The R-squared value.

ForestMensuration.regression — Method

Perform regression analysis.

Arguments

y::S: Symbol representing the dependent variable.

x::S: Symbol representing the independent variable.

data::AbstractDataFrame: Data frame containing the data.

q::S...: Symbols representing additional terms for the model.

best::Union{Bool, Int}: Whether to return the best model(s) based on RMSE.

effect::S: Type of effect to consider (:additive or :interactive).

Returns

Vector{FittedLinearModel}: Vector of fitted models.

ForestMensuration.residuals — Method

Calculate residuals of a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Vector{Float64}: The residuals.

ForestMensuration.simple_casual_sampling — Method

Performs simple random sampling for forest inventory with specified plot area and total area.

Arguments

volume::Vector{<:Real}: Volume measurements.

plot_area::Real: Area of each plot.

total_area::Real: Total area.

e::Real=10: Desired error margin (default is 10%).

α::Real=0.95: Confidence level (default is 95%).

lg::Symbol=:pt: Language for the report (default is Portuguese).

Returns

DataFrame: Inventory report.

ForestMensuration.simple_casual_sampling — Method

Performs simple random sampling for forest inventory with total area.

Arguments

volume::Vector{<:Real}: Volume measurements.

total_area::Real: Total area.

e::Real=10: Desired error margin (default is 10%).

α::Real=0.95: Confidence level (default is 95%).

lg::Symbol=:pt: Language for the report (default is Portuguese).

Returns

DataFrame: Inventory report.

ForestMensuration.stderror — Method

Calculate the standard error of the coefficients of a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Vector{Float64}: The standard errors.

ForestMensuration.syx — Method

Calculate the standard error of the estimate for a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Float64: The standard error of the estimate.

ForestMensuration.syx_in_percentage — Method

Calculate the standard error of the estimate as a percentage for a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Float64: The standard error of the estimate as a percentage.

StatsAPI.aic — Method

Calculate the Akaike Information Criterion (AIC) for a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Float64: The AIC value.

StatsAPI.aicc — Method

Calculate the corrected Akaike Information Criterion (AICc) for a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Float64: The AICc value.

StatsAPI.bic — Method

Calculate the Bayesian Information Criterion (BIC) for a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Float64: The BIC value.

StatsAPI.coef — Method

Get coefficients of a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Vector{Float64}: The coefficients.

StatsAPI.deviance — Method

Calculate the deviance for a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Float64: The deviance.

StatsAPI.dof — Method

Get degrees of freedom of a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Int: The degrees of freedom.

StatsAPI.dof_residual — Method

Get degrees of freedom for residuals of a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Int: The degrees of freedom for residuals.

StatsAPI.modelmatrix — Method

Get the model matrix of a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Matrix{Float64}: The model matrix.

StatsAPI.nobs — Method

Get the number of observations used in fitting the model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Int: The number of observations.

StatsAPI.nulldeviance — Method

Calculate the null deviance for a fitted linear model.

Arguments

fitted_model::FittedLinearModel: The fitted linear model.

Returns

Float64: The null deviance.

StatsAPI.nulldeviance — Method

Calculate the null deviance for a vector of values.

Arguments

x::Vector{<:Real}: The vector of values.

Returns

Float64: The null deviance.