Reference

Contents

Reference

Contents

Index

Index

ForestMensuration.ForestMensuration

ForestMensuration.mySchema

ForestMensuration.β0

ForestMensuration.CubingMethod

ForestMensuration.Huber

ForestMensuration.MixTerm

ForestMensuration.Newton

ForestMensuration.Smalian

ForestMensuration._amplitude

ForestMensuration._calculate_delta

ForestMensuration._class_breadth

ForestMensuration._class_center

ForestMensuration._dependent_variable

ForestMensuration._dependent_variable

ForestMensuration._finite_population

ForestMensuration._finite_sample

ForestMensuration._independent_variable

ForestMensuration._infinite_population

ForestMensuration._infinite_sample

ForestMensuration._inventory_report

ForestMensuration._simple_frequency

ForestMensuration._sturges

ForestMensuration.artificial_form_factor

ForestMensuration.bark_factor

ForestMensuration.bole_volume

ForestMensuration.cone_volume

ForestMensuration.criteria_selection

ForestMensuration.criteria_table

ForestMensuration.cubage

ForestMensuration.cubage

ForestMensuration.cubage

ForestMensuration.cubage

ForestMensuration.cylinder_volume

ForestMensuration.diametric_table

ForestMensuration.diametric_table

ForestMensuration.diametric_table

ForestMensuration.diametric_table

ForestMensuration.frequency_table

ForestMensuration.frequency_table

ForestMensuration.frequency_table

ForestMensuration.frequency_table

ForestMensuration.hdom_classification

ForestMensuration.natural_form_factor

ForestMensuration.plot_regression

ForestMensuration.prediction

ForestMensuration.prediction!

ForestMensuration.quotient_form

ForestMensuration.regression

ForestMensuration.simple_casual_sampling

ForestMensuration.simple_casual_sampling

ForestMensuration.site_classification

ForestMensuration.site_classification

ForestMensuration.site_table

ForestMensuration.ForestMensuration — Module

Description:

ForestMensuration.jl is a Julia package that provides a comprehensive set of functions for performing dendrometric and forest inventory calculations. The package emphasizes ease of use, making it straightforward to:

Conduct regressions: Linear models can be fitted with ease.

Calculate tree and stand volume: Various methods are supported, including Huber, Smalian, and Newton.

Perform forest inventories: Implementations of simple random sampling are available.

Create frequency tables: Analyze the distribution of dendrometric variables.

Facilitates the analysis of dendrometric and forest data. Performs complex calculations with simple commands. Offers a user-friendly and intuitive interface.

ForestMensuration.MixTerm — Type

Union type representing a mixed term, which can be a single AbstractTerm or a tuple of AbstractTerms.

ForestMensuration.mySchema — Constant

Represents a default schema used for modeling operations.

ForestMensuration.β0 — Constant

Represents an intercept term for linear models.

ForestMensuration.CubingMethod — Type

Abstract type representing a method for cubing (calculating volume).

ForestMensuration.Huber — Type

Huber Method: The Huber method measures the diameter or circumference at the midpoint of the section, and the volume is determined by:

v = v0 + Σi=1:n(vi) + vt

vi = gi * li

Where:

v0 = volume of the stump;

vi = volume of intermediate sections;

vt = volume of the cone;

g = basal area;

l = length.

ForestMensuration.Newton — Type

Newton Method: The Newton method involves measuring at 3 positions along each section (at the ends and in the middle of the logs). Therefore, it is a more laborious method than the others, but the estimated volume will be more accurate.

v = v0 + Σi=1:n(vi) + vt

vi = (gi + gm + gi+1)/2 * li

Where:

v0 = volume of the stump;

vi = volume of intermediate sections;

vt = volume of the cone;

g = basal area;

gm = basal area at the midpoint of the section;

l = length.

ForestMensuration.Smalian — Type

Smalian Method: The Smalian method measures diameters or circumferences at the ends of each section and calculates the total volume by:

Vt = v0 + Σi=1:n(vi) + vt

v0 = g0 * l0

vi = (gi + gi+1)/2 * li

vt = (1/3) * gn * ln

Where:

v0 = volume of the stump;

vi = volume of intermediate sections;

vt = volume of the cone;

g = basal area;

l = length.

ForestMensuration._amplitude — Method

Calculates the amplitude (range) of a vector of values.

Arguments

x::Vector: The vector of values.

Returns

Real: The amplitude (range).

ForestMensuration._calculate_delta — Method

Calculate the delta (difference) between the model matrices for age and index age.

Arguments

model::StatsModels.TableRegressionModel: The fitted regression model.

data_age::AbstractDataFrame: The data frame containing the current age data.

index_age::Real: The index age for which the delta is calculated.

Returns

Δ::Vector{Real}: The delta for each observation.

ForestMensuration._class_breadth — Method

Calculates the class breadth (width) for a given amplitude and number of classes.

Arguments

h::Real: The amplitude.

k::Int: The number of classes.

Returns

Real: The class breadth (width).

ForestMensuration._class_center — Method

Calculates the class center for a given value and class width.

Arguments

x::Real: The value.

hi::Real: The class width.

Returns

Real: The class center.

ForestMensuration._dependent_variable — Method

Generates a list of transformed dependent variable terms.

Arguments

y_term::AbstractTerm: The dependent variable term.

x_term::AbstractTerm: The independent variable term.

Returns

Vector{AbstractTerm}: A list of transformed dependent variable terms.

ForestMensuration._dependent_variable — Method

Generates a list of transformed dependent variable term.

Arguments

y_term::AbstractTerm: The dependent variable term.

Returns

Vector{AbstractTerm}: A list of transformed dependent variable term.

ForestMensuration._finite_population — Method

Calculates the required sample size for a finite population using an iterative approach with a maximum iteration limit to prevent infinite loops.

Arguments

cv::Real: Coefficient of variation.

ni::Real: Initial sample size.

N::Int: Population size.

e::Real: Desired error margin.

α::Real: Confidence level.

max_iterations::Int=1000: Maximum number of iterations (default is 1000).

Returns

Int64: The required sample size for a finite population.

ForestMensuration._finite_sample — Method

Calculates the sample size for a finite population.

Arguments

t::Real: T-value for the given confidence level.

cv::Real: Coefficient of variation.

e::Real: Desired error margin.

N::Int64: Population size.

Returns

Float64: The sample size for a finite population.

ForestMensuration._independent_variable — Method

Generates a list of transformed independent variable terms and their interactions.

Arguments

x_term::AbstractTerm: The independent variable term.

Returns

Vector{MixTerm}: A list of transformed independent variable terms and their interactions.

ForestMensuration._infinite_population — Method

Calculates the required sample size for an infinite population using an iterative approach with a maximum iteration limit to prevent infinite loops.

Arguments

cv::Real: Coefficient of variation.

ni::Real: Initial sample size.

e::Real: Desired error margin.

α::Real: Confidence level.

max_iterations::Int=1000: Maximum number of iterations (default is 1000).

Returns

Int64: The required sample size for an infinite population.

ForestMensuration._infinite_sample — Method

Calculates the sample size for an infinite population.

Arguments

t::Real: T-value for the given confidence level.

cv::Real: Coefficient of variation.

e::Real: Desired error margin.

Returns

Float64: The sample size for an infinite population.

ForestMensuration._inventory_report — Method

Generate an inventory report.

Arguments

x̅::F: Mean volume per plot.

cv::F: Coefficient of variation.

Ttab::F: T-value for the confidence level.

s²x̅::F: Mean variance.

sx̅::F: Standard error.

plot_area::Union{I, F, S}: Area of the plot.

f::F: Sampling fraction.

population::String: Population type ("infinite" or "finite").

n::Union{I, T}: Number of measured plots.

required_plots::Union{I, T}: Number of required plots.

missing_plots::Union{I, T}: Number of missing plots.

N::I: Total number of plots.

lg::Symbol: Language for the report (":pt" for Portuguese, ":en" for English).

Returns

DataFrame: The inventory report.

ForestMensuration._simple_frequency — Method

Calculates the simple frequency of unique values in a vector.

Arguments

x::Vector: The vector of values.

Returns

Vector: The frequency of each unique value.

ForestMensuration._sturges — Method

Calculates the number of classes using Sturges' formula.

Arguments

n::Int: The number of observations.

Returns

Int: The number of classes.

ForestMensuration.artificial_form_factor — Method

Artificial Form Factor (aff): For the calculation of the artificial form factor, the volume of the reference cylinder will have a diameter equal to the tree's DBH.

aff = Rigorous Vol / Cylinder Vol 1.3

Where:

Rigorous Vol = total volume determined by one of the methods: Smalian, Huber, or Newton;

Cylinder Vol 1.3 = volume of a cylinder with height and diameter equal to the total height and DBH of the tree.

Arguments

vt::Real: The total volume of the tree.

ht::Real: The total height of the tree.

dbh::Real: The diameter at breast height of the tree.

Returns

Float64: The artificial form factor.

ForestMensuration.bark_factor — Method

Calculates the bark factor, used to estimate the volume without bark.

The bark factor is used to estimate the volume without bark by considering the ratio of bark thickness to total diameter.

Arguments

d::Vector{<:Real}: Vector of diameters.

e::Vector{<:Real}: Vector of bark thicknesses in millimeters.

Returns

Float64: The bark factor, which represents the proportion of the diameter without bark.

Example

Calculate bark factor

k = bark_factor(d, e)

Calculate total volume without bark

totalvolumewithoutbark = totalvolume * k^2

ForestMensuration.bole_volume — Method

Calculate tree bole volume cubic meters using Smalian, Newton, or Huber methods. The methods involve dividing the tree trunk into n sections (logs). In each section, diameters and lengths are measured at positions that vary according to the technique employed.

Arguments

method::Type{<:CubingMethod}: The method used for cubing (Smalian, Huber, or Newton).

h::Vector{<:Real}: Vector of heights.

d::Vector{<:Real}: Vector of diameters.

Returns

Float64: The volume of the bole in cubic meters.

ForestMensuration.cone_volume — Method

Calculates the volume of a cone, used to estimate the final portion (vn) of the tree, typically considered to have a conical shape.

Arguments

h::Real: The height of the cone in meters.

d::Real: The diameter of the cone in centimeters.

Returns

Float64: The volume of the cone in cubic meters.

Throws

ArgumentError: If h or d is less than or equal to zero, as height and diameter must be positive values.

ForestMensuration.criteria_selection — Method

The criteria_selection function evaluates and ranks a vector of regression models based on specified criteria, returning the best model according to the combined ranking.

Parameters:

model::Vector{<:TableRegressionModel}: A vector of linear regression models to be evaluated and compared.

criteria::Symbol...: A variable number of symbols representing the evaluation criteria to include. Possible values include:

:adjr2: Adjusted R², a measure of the model's explanatory power, adjusted for the number of predictors.

:syx: Standard error of the estimate as a percentage of the mean of y.

:rmse: Root Mean Squared Error, indicating the average magnitude of residuals.

:mae: Mean Absolute Error, another accuracy measure based on average absolute residuals.

:aic: Akaike Information Criterion, balancing goodness of fit with model complexity.

:significance: Evaluates whether model coefficients are statistically significant.

:normality: Assesses the normality of residuals, an assumption in linear regression.

:homoscedasticity: Checks for constant variance in residuals, another key regression assumption.

If no criteria are specified, the function will use all available criteria by default.

Returns:

TableRegressionModel: The best model based on the combined ranking of the specified criteria.

ForestMensuration.criteria_table — Method

The criteria_table function evaluates and ranks multiple regression models based on specified criteria. It generates a comprehensive table of performance metrics for each model, calculates ranks for these metrics, and combines them into a final score. The function allows for flexible selection of evaluation criteria and can return either all models or only the top models based on the combined ranking.

Parameters:

model: The regression model(s) to be evaluated and compared. This parameter can accept:

Single Linear Regression Model (TableRegressionModel): Evaluates a single linear regression model.

Vector of Linear Regression Models (Vector{<:TableRegressionModel}): Evaluates and compares multiple linear regression models.

criteria::Symbol...: A variable number of symbols representing the evaluation criteria to include. Possible values include:

:adjr2: Adjusted R², a measure of the model's explanatory power, adjusted for the number of predictors.

:syx: Standard error of the estimate (Syx) expressed as a percentage of the mean of the dependent variable (y), indicating the precision of the model's predictions.

:rmse: Root Mean Squared Error, indicating the average magnitude of residuals.

:mae: Mean Absolute Error, another accuracy measure based on average absolute residuals.

:aic: Akaike Information Criterion, balancing goodness of fit with model complexity.

:normality: Assesses the normality of residuals using the Kolmogorov-Smirnov test, ensuring that residuals follow a normal distribution.

:significance: Evaluates whether model coefficients are statistically significant.

If no criteria are specified, the function will use all available criteria by default.

best::Union{Bool, Int}=10: Specifies the number of top models to return based on the combined ranking.

false: Returns the full table with all models ranked.

Integer value: If less than the total number of models, returns only the top best models.

Returns:

DataFrame: A sorted table with the evaluated models and their respective metrics. Includes a combined rank based on the selected criteria.

Examples:

Single Model: criteria_table(model, :adjr2, :rmse)

Vector of Models: criteria_table([model1, model2], :aic, :mae)

ForestMensuration.cubage — Function

Calculate tree cubage using grouped data from a DataFrame. The methods involve dividing the tree trunk into n sections (logs). In each section, diameters and lengths are measured at positions that vary according to the technique employed.

Arguments

method::Type{<:CubingMethod}: The method used for cubing (Smalian, Huber, or Newton).

tree::Symbol: The symbol representing the tree identifier.

h::Symbol: The symbol representing the heights in the DataFrame.

d::Symbol: The symbol representing the diameters in the DataFrame.

data::AbstractDataFrame: The DataFrame containing the tree data.

d_limit::Union{Float64, Nothing}: Comercial diameter limit to be used in calculations (optional).

dbh::Float64: Diameter at breast height (default is 1.3 meters).

Returns

DataFrame: A DataFrame with the calculated volumes and form factors for each tree.

ForestMensuration.cubage — Function

Calculate tree cubage using Smalian, Newton, or Huber methods. The methods involve dividing the tree trunk into n sections (logs). In each section, diameters and lengths are measured at positions that vary according to the technique employed. Thus, the volume of the sections and the total volume are determined by summing the volume of the sections. Determination can be carried out on felled trees or standing trees using equipment such as the Bitterlich relascope.

Arguments

method::Type{<:CubingMethod}: The method used for cubing (Smalian, Huber, or Newton).

h::Vector{<:Real}: Vector of heights.

d::Vector{<:Real}: Vector of diameters.

d_limit::Union{Float64, Nothing}: Comercial diameter limit to be used in calculations (optional).

dbh::Float64: Diameter at breast height (default is 1.3 meters).

Returns

DataFrame: A DataFrame with the calculated volumes and form factors.

ForestMensuration.cubage — Function

Calculate tree cubage including bark factor. The methods involve dividing the tree trunk into n sections (logs). In each section, diameters and lengths are measured at positions that vary according to the technique employed.

Arguments

method::Type{<:CubingMethod}: The method used for cubing (Smalian, Huber, or Newton).

h::Vector{<:Real}: Vector of heights.

d::Vector{<:Real}: Vector of diameters.

e::Vector{<:Real}: Vector of bark thicknesses.

d_limit::Union{Float64, Nothing}: Comercial diameter limit to be used in calculations (optional).

dbh::Float64: Diameter at breast height (default is 1.3 meters).

Returns

DataFrame: A DataFrame with the calculated volumes, form factors, and bark-adjusted volumes.

ForestMensuration.cubage — Function

Calculate tree cubage including bark factor using grouped data from a DataFrame. The methods involve dividing the tree trunk into n sections (logs). In each section, diameters and lengths are measured at positions that vary according to the technique employed.

Arguments

method::Type{<:CubingMethod}: The method used for cubing (Smalian, Huber, or Newton).

tree::Symbol: The symbol representing the tree identifier.

h::Symbol: The symbol representing the heights in the DataFrame.

d::Symbol: The symbol representing the diameters in the DataFrame.

e::Symbol: The symbol representing the bark thicknesses in the DataFrame.

data::AbstractDataFrame: The DataFrame containing the tree data.

d_limit::Union{Float64, Nothing}: Comercial diameter limit to be used in calculations (optional).

dbh::Float64: Diameter at breast height (default is 1.3 meters).

Returns

DataFrame: A DataFrame with the calculated volumes, form factors, and bark-adjusted volumes for each tree.

ForestMensuration.cylinder_volume — Method

Calculates the volume of a cylinder, used to estimate the volume (v0) of the tree stump remaining after clear-cutting.

Arguments

h::Real: The height of the cylinder in meters.

d::Real: The diameter of the cylinder in centimeters.

Returns

Float64: The volume of the cylinder in cubic meters.

Throws

ArgumentError: If h or d is less than or equal to zero, as height and diameter must be positive values.

ForestMensuration.diametric_table — Method

Creates a diametric table for a vector of values given a class width and plot area.

Arguments

x::Vector{<:Real}: The vector of values.

hi::Real: The class width.

plot_area::Real=1.0: The plot area in hectares (default is 1.0).

Returns

DataFrame: A DataFrame containing the diametric table.

ForestMensuration.diametric_table — Method

Creates a diametric table for a vector of values.

Arguments

x::Vector{<:Real}: The vector of values.

plot_area::Real=1.0: The plot area in hectares (default is 1.0).

Returns

DataFrame: A DataFrame containing the diametric table.

ForestMensuration.diametric_table — Method

Creates a diametric table for grouped data in a DataFrame.

Arguments

g::S: The symbol representing the grouping variable.

x::S: The symbol representing the values variable.

data::AbstractDataFrame: The DataFrame containing the data.

plot_area::Real=1.0: The plot area in hectares (default is 1.0).

Returns

DataFrame: A DataFrame containing the diametric table for each group.

ForestMensuration.diametric_table — Method

Creates a diametric table for grouped data in a DataFrame with a specified class width.

Arguments

g::S: The symbol representing the grouping variable.

x::S: The symbol representing the values variable.

hi::Real: The class width.

data::AbstractDataFrame: The DataFrame containing the data.

plot_area::Real=1.0: The plot area in hectares (default is 1.0).

Returns

DataFrame: A DataFrame containing the diametric table for each group.

ForestMensuration.frequency_table — Method

Creates a frequency table for a vector of values given a class width.

Arguments

x::Vector{<:Real}: The vector of values.

hi::Real: The class width.

Returns

DataFrame: A DataFrame containing the frequency table.

ForestMensuration.frequency_table — Method

Creates a frequency table for a vector of values.

Arguments

x::Vector{<:Real}: The vector of values.

Returns

DataFrame: A DataFrame containing the frequency table.

ForestMensuration.frequency_table — Method

Creates a frequency table for grouped data in a DataFrame.

Arguments

g::S: The symbol representing the grouping variable.

x::S: The symbol representing the values variable.

data::AbstractDataFrame: The DataFrame containing the data.

Returns

DataFrame: A DataFrame containing the frequency table for each group.

ForestMensuration.frequency_table — Method

Creates a frequency table for grouped data in a DataFrame with a specified class width.

Arguments

g::S: The symbol representing the grouping variable.

x::S: The symbol representing the values variable.

hi::Real: The class width.

data::AbstractDataFrame: The DataFrame containing the data.

Returns

DataFrame: A DataFrame containing the frequency table for each group.

ForestMensuration.hdom_classification — Method

Calculate the dominant height given the site classification, fitted model, and index age.

Arguments

model::StatsModels.TableRegressionModel: The fitted regression model.

data_age::AbstractDataFrame: The data frame containing the current age data.

index_age::Real: The index age for the calculation.

site::Vector{<:Real}: The site classification values.

Returns

hdom::Vector{Real}: The dominant height for each observation.

ForestMensuration.natural_form_factor — Method

Natural Form Factor (nff): For the calculation of the natural form factor, the volume of the reference cylinder will have a diameter equal to the diameter taken at 1/10 of the total height.

f0.1h = Rigorous Vol / Cylinder Vol 0.1

Where:

Rigorous Vol = total volume determined by one of the methods: Smalian, Huber, or Newton;

Cylinder Vol 0.1 = volume of a cylinder with height equal to the total height of the tree and diameter taken at 1/10 of the total height.

Interpolate diameter at a given height using linear interpolation.

Arguments

vt::Real: The total volume of the tree.

ht::Real: The total height of the tree.

h::Vector{<:Real}: Vector of heights.

d::Vector{<:Real}: Vector of diameters.

Returns

Float64: The natural form factor.

ForestMensuration.plot_regression — Method

The plot_regression function generates four essential diagnostic plots to analyze the performance and validity of a linear regression model in Julia. These plots help in assessing the goodness-of-fit, the distribution of residuals, and the assumptions underlying linear regression, such as homoscedasticity and normality. The function is particularly useful for ensuring that the regression model meets the assumptions required for reliable inference and prediction.

Parameters:

model::TableRegressionModel: The fitted linear regression model. This model is analyzed to generate the diagnostic plots.

Functionality:

Observed vs Fitted Values: This scatter plot compares the observed data points with the values predicted by the model. It is useful for visually assessing the accuracy of the model's predictions.

Residuals vs Fitted Values: This plot shows the residuals (errors) against the fitted values. Ideally, residuals should be randomly scattered around zero without any apparent pattern, indicating a good model fit.

Histogram of Residuals: This histogram displays the distribution of residuals. A normal distribution of residuals is one of the key assumptions of linear regression, and this plot helps verify it.

Normal Q-Q Plot: This plot compares the quantiles of the residuals to the quantiles of a normal distribution. It is used to check the assumption of normality in the residuals.

The function automatically applies color schemes and adjusts plot aesthetics to enhance the clarity of the visualizations. The plots are combined into a single layout for easy comparison and interpretation. Additionally, the function adjusts dynamically based on the number of groups in the data, making it versatile for different datasets and models.

Examples:

# Generate diagnostic plots for a fitted regression model
plots = plot_regression(model)

ForestMensuration.prediction! — Method

The prediction! function computes predictions from a regression model and adds these predictions directly to the provided data frame as new columns. It is particularly useful in forest inventory data where not all trees have been measured for a specific variable, allowing the model to estimate these missing values.

Parameters:

model: A single linear regression model.

data: The data frame (AbstractDataFrame) containing the input data. The function will add new columns to this data frame.

Functionality:

Predicted Values Column (_predict): The function calculates the predicted values for the dependent variable (y) based on the input model and appends these values as a new column in the data frame with the suffix _predict.

Real or Estimated Values Column (_real): The function also creates a _real column where the actual measured values of y are preserved if they exist. For observations where y is missing (or set to 0.0), the predicted value from the model is used instead.

This setup is ideal for forest inventory datasets where certain tree attributes (like height or volume) might not be measured for every tree, and predictions need to be filled in for these gaps.

Examples:

# Apply predictions to a data frame
prediction!(model, forest_inventory_data)

ForestMensuration.prediction — Method

The prediction function family provides a versatile way to generate predictions from regression models, supporting both individual and grouped models. It handles predictions on the original scale even if the dependent variable (y) has been transformed (e.g., log(y)), ensuring that any transformations applied during model fitting are correctly reversed, including the application of Meyer correction factors for logarithmic transformations.

Parameters:

model: A single linear regression model.

data: The input data for which predictions are needed. This parameter can be:

AbstractDataFrame: A data frame containing the input data.

DataFrameRow: A single row of data from a data frame.

Returns:

Vector{<:Real} or Vector{Union{Missing, <:Real}}: The predicted values on the original scale of y, adjusted for any transformations and corrected using the Meyer factor for logarithmic transformations.

Key Features:

Handles Transformed Dependent Variables: If the dependent variable was transformed (e.g., using log transformations), the function correctly inverts the transformation to return predictions on the original scale.

Applies Meyer Correction Factor: For models using logarithmic transformations, the Meyer correction factor is applied to the predictions to correct for the bias introduced by the log transformation.

Supports Grouped Models: When using grouped models, the function automatically selects and applies the correct model for each data point based on its group membership.

Examples:

Single Model Prediction:
y_pred = prediction(model, data)

ForestMensuration.quotient_form — Method

Form Quotient (qf): The natural decrease in diameter along the trunk defines the so-called form quotient, which is a ratio between diameters. An example of a form quotient is the Schiffel form quotient, given by:

Q = D(1/2H) / DBH

Where:

Q < 1

D(1/2H) = diameter measured at half the total height of the tree.

Similar to the form factor, the volume of a tree, with or without bark, can be obtained by multiplying the volume of a cylinder by the average form quotient, suitable for the species and the desired volume to be estimated.

Arguments

ht::Real: The total height of the tree.

dbh::Real: The diameter at breast height of the tree.

h::Vector{<:Real}: Vector of heights.

d::Vector{<:Real}: Vector of diameters.

Returns

Float64: The form quotient.

ForestMensuration.regression — Method

The regression function in Julia automatically generates and evaluates multiple simple regression models based on the provided data, including both continuous and categorical variables. This function significantly expands the traditional analysis typically applied in forest biometrics, such as the relationship between tree height and diameter at breast height (DBH), by automatically generating and evaluating 240 unique combinations of dependent and independent variable transformations.

Parameters:

y::Symbol: The dependent variable (response variable) for the regression model. Typically represents a continuous measure such as tree height or volume in forestry studies.

x::Symbol: The independent variable (predictor variable) used to explain variations in the dependent variable. Often represents a measure like diameter at breast height (DBH) in forestry.

data::AbstractDataFrame: The dataset containing the variables for regression. The data frame must include all variables specified in y, x, and q, and it will automatically remove any rows with missing values before performing the regression.

q::Symbol... (optional): A variable number of symbols representing additional categorical variables (qualitative factors) to include in the regression model. These are treated as factors and can influence the model differently based on their levels.

Functionality:

Dependent Variable Transformations: The function generates multiple transformations of the dependent variable (y), including logarithmic and inverse functions. Specifically, the following transformations are created:

log(y): Logarithm of y.

log_minus(y): Logarithm of y - 1.3.

log1p(y): Logarithm of 1 + y.

1/y: Inverse of y.

1/(y - 1.3): Inverse of y - 1.3.

1/√y: Inverse of the square root of y.

1/√(y - 1.3): Inverse of the square root of y - 1.3.

x/√y: Ratio of x to the square root of y.

x/√(y - 1.3): Ratio of x to the square root of y - 1.3.

x²/y: Square of x divided by y.

x²/(y - 1.3): Square of x divided by y - 1.3.

Independent Variable Transformations: Similarly, multiple transformations of the independent variable (x) are created:

x²: Square of x.

log(x): Logarithm of x.

log(x)²: Square of the logarithm of x.

1/x: Inverse of x.

1/x²: Inverse of Square of x.

Combined Model Formulations: The function creates a total of 240 combinations of these transformations by pairing the various forms of y and x. For example, the function considers combinations such as:

y ~ x + x²

y ~ x + log(x)

log(y) ~ log(x) + 1/x

(x / y) ~ x + log(x)²

And many more.

This comprehensive set of models extends beyond the typical scope of forest biometrics, where usually only a few standard models (around five) are tested. By automatically exploring a wide array of potential relationships, this method allows for a more thorough investigation of the possible functional forms that describe the relationship between tree height and DBH or other relevant variables.

Standard Regression: The function performs a regression analysis by automatically generating a wide array of possible models. It creates multiple transformations of the dependent and independent variables, combining them into various model forms. The results can be evaluated, and the best models can be selected based on criteria such as adjusted R², RMSE, AIC, and more, using the criteria_table function.

Qualitative Variables: The function allows the inclusion of categorical variables (q) in the regression model. These variables are automatically treated as factors and can be used to capture variations in the dependent variable that are related to these qualitative factors.

Applications:

This method is ideal for forestry researchers and practitioners who want to ensure they are not overlooking any potentially significant models by expanding their analysis to include a broader range of possible relationships between the variables.

Examples:

# Perform standard regression without grouping
models = regression(:height, :diameter, data)

# View the top models based on a specific criteria
best_models = criteria_table(models, :adjr2, :rmse)

ForestMensuration.simple_casual_sampling — Method

Performs simple random sampling for forest inventory with specified plot area and total area.

Arguments

volume::Vector{<:Real}: Volume measurements.

plot_area::Real: Area of each plot.

total_area::Real: Total area.

e::Real=10: Desired error margin (default is 10%).

α::Real=0.95: Confidence level (default is 95%).

lg::Symbol=:pt: Language for the report (default is Portuguese).

Returns

DataFrame: Inventory report.

ForestMensuration.simple_casual_sampling — Method

Performs simple random sampling for forest inventory with total area.

Arguments

volume::Vector{<:Real}: Volume measurements.

total_area::Real: Total area.

e::Real=10: Desired error margin (default is 10%).

α::Real=0.95: Confidence level (default is 95%).

lg::Symbol=:pt: Language for the report (":pt" for Portuguese - default, ":en" for English).

Returns

DataFrame: Inventory report.

ForestMensuration.site_classification — Method

Calculate the site classification given the fitted model and index age.

Arguments

model::StatsModels.TableRegressionModel: The fitted regression model.

data_age::AbstractDataFrame: The data frame containing the current age data.

index_age::Real: The index age for site classification.

Returns

site::Vector{Real}: The site classification for each observation.

ForestMensuration.site_classification — Method

Calculate the site classification given the fitted model and index age.

Arguments

model: The fitted regression model.

index_age::Real: The index age for site classification.

Returns

site_classification::Vector{Real}: The site classification for each observation.

ForestMensuration.site_table — Method

Calculate the site table given a fitted model, index age, and height increment.

Arguments

model::StatsModels.TableRegressionModel: The fitted regression model.

index_age::Real: The index age for site classification.

hi::Real: The height increment for site classification.

Returns

SiteAnalysis: A struct containing site table and site plot.