Reference

Contents

Reference

Contents

Index

Index

ForestMensuration.ForestMensuration

ForestMensuration.Huber

ForestMensuration.Newton

ForestMensuration.Smalian

ForestMensuration.artificial_form_factor

ForestMensuration.bark_factor

ForestMensuration.bole_volume

ForestMensuration.cone_volume

ForestMensuration.cubage

ForestMensuration.cylinder_volume

ForestMensuration.natural_form_factor

ForestMensuration.quotient_form

ForestMensuration.syx

ForestMensuration.ForestMensuration — Module

Description:

ForestMensuration.jl is a Julia package that provides a comprehensive set of functions for performing dendrometric/biometric and forest inventory calculations. The package emphasizes ease of use, making it straightforward to:

Conduct regressions: Linear, nonlinear, and mixed-effects models can be fitted with ease. Calculate tree and stand volume: Various methods are supported, including Huber, Smalian, and Newton. Perform forest inventories: Implementations of simple random and stratified sampling are available. Create frequency tables: Analyze the distribution of dendrometric variables. Work with probability distributions: Calculate PDF and CDF for various distributions.

Facilitates the analysis of dendrometric and forest data. Performs complex calculations with simple commands. Offers a user-friendly and intuitive interface.

ForestMensuration.Huber — Type

Huber Method: The Huber method measures the diameter or circumference at the midpoint of the section, and the volume is determined by:

v = v0 + Σi=1:n(vi) + vt

vi = gi * li

Where:

v0 = volume of the stump;

vi = volume of intermediate sections;

vt = volume of the cone;

g = basal area;

l = length.

ForestMensuration.Newton — Type

Newton Method: The Newton method involves measuring at 3 positions along each section (at the ends and in the middle of the logs). Therefore, it is a more laborious method than the others, but the estimated volume will be more accurate.

v = v0 + Σi=1:n(vi) + vt

vi = (gi + gm + gi+1)/2 * li

Where:

v0 = volume of the stump;

vi = volume of intermediate sections;

vt = volume of the cone;

g = basal area;

gm = basal area at the midpoint of the section;

l = length.

ForestMensuration.Smalian — Type

Smalian Method: The Smalian method measures diameters or circumferences at the ends of each section and calculates the total volume by:

Vt = v0 + Σi=1:n(vi) + vt

v0 = g0 * l0

vi = (gi+gi+1)/2 * li

vt = (1/3) * gn * ln

Where:

v0 = volume of the stump;

vi = volume of intermediate sections;

vt = volume of the cone;

g = basal area;

l = length.

ForestMensuration.artificial_form_factor — Method

Artificial Form Factor (aff): For the calculation of the artificial form factor, the volume of the reference cylinder will have a diameter equal to the tree's DBH.

aff = Rigorous Vol / Cylinder Vol 1,3

Where:

Rigorous Vol = total volume determined by one of the methods: Smalian, Huber, or Newton;

Cylinder Vol 1,3 = volume of a cylinder with height and diameter equal to the total height and DBH of the tree.

ForestMensuration.bark_factor — Method

Calculates the bark factor, used to estimate the volume without bark.

The bark factor is used to estimate the volume without bark by considering the ratio of bark thickness to total diameter.

Arguments

d::Vector{<:Real}: Vector of diameters.

e::Vector{<:Real}: Vector of bark thicknesses in millimeters.

Returns

Float64: The bark factor.

Example

Calculate bark factor

bk = bark_factor(d, e)

Calculate total volume without bark

totalvolumewithoutbark = totalvolume * bk^2

ForestMensuration.bole_volume — Method

Calculate tree bole volume cubic meters using Smalian, Newton, or Huber methods. The methods involve dividing the tree trunk into n sections (logs). In each section, diameters and lengths are measured at positions that vary according to the technique employed.

ForestMensuration.cone_volume — Method

Calculates the volume of a cone, used to estimate the final portion (vn) of the tree, typically considered to have a conical shape.

Arguments

h::Real: The height of the cone in meters.

d::Real: The diameter of the cone in centimeters.

Returns

Float64: The volume of the cone in cubic meters.

ForestMensuration.cubage — Method

Calculate tree cubage using Smalian, Newton, or Huber methods. The methods involve dividing the tree trunk into n sections (logs). In each section, diameters and lengths are measured at positions that vary according to the technique employed. Thus, the volume of the sections and the total volume are determined by summing the volume of the sections. Determination can be carried out on felled trees or standing trees using equipment such as the Bitterlich relascope.

ForestMensuration.cylinder_volume — Method

Calculates the volume of a cylinder, used to estimate the volume (v0) of the tree stump remaining after clear-cutting.

Arguments

h::Real: The height of the cylinder in meters.

d::Real: The diameter of the cylinder in centimeters.

Returns

Float64: The volume of the cylinder.

ForestMensuration.natural_form_factor — Method

Natural Form Factor (nff): For the calculation of the natural form factor, the volume of the reference cylinder will have a diameter equal to the diameter taken at 1/10 of the total height.

f0,1h = Rigorous Vol / Cylinder Vol 0,1

Where:

Rigorous Vol = total volume determined by one of the methods: Smalian, Huber, or Newton;

Cylinder Vol 0,1 = volume of a cylinder with height equal to the total height of the tree and diameter taken at 1/10 of the total height.

Interpolate diameter at a given height using linear interpolation.

ForestMensuration.quotient_form — Method

Form Quotient (qf): The natural decrease in diameter along the trunk defines the so-called form quotient, which is a ratio between diameters. An example of a form quotient is the Schiffel form quotient, given by:

Q = D(1/2H) / DBH, Where

Q < 1

D(1/2H) = diameter measured at half the total height of the tree.

Similar to the form factor, the volume of a tree, with or without bark, can be obtained by multiplying the volume of a cylinder by the average form quotient, suitable for the species and the desired volume to be estimated.

ForestMensuration.syx — Method

The standard error of the estimate is a measure of the accuracy of predictions.